refactor: move message-request and reaction-picker handlers into handlers/keys (#494)
Next slice of the handlers/ extraction. handle_message_request_key and
handle_reaction_picker_key move from inline impl App methods into handlers::keys
as free functions over &mut App. Both were private and are called from the
overlay dispatcher and a few tests, so app.rs keeps private method shims that
delegate (the dispatcher and tests are unchanged).

prepare_reaction_send is promoted to pub(crate) so the moved reaction handler
can call it; prepare_reaction_send_emoji stays private (only used within app.rs).
Behavior unchanged.

No App fields changed; field-count ratchet unaffected.
